+++ dbus-serialbattery/dbus-serialbattery.py
@@ -365,12 +365,14 @@ def main():
         """
         from bms.daly_can import Daly_Can
         from bms.jkbms_can import Jkbms_Can
+        from bms.rv_c_can import RV_C_Can
         from bms.ubms_can import Ubms_Can
 
         # only try CAN BMS on CAN port
         supported_bms_types = [
             {"bms": Daly_Can},
             {"bms": Jkbms_Can},
+            {"bms": RV_C_Can},
             {"bms": Ubms_Can},
         ]